Add AST class generator and a todo file
Through sweat and pain, here comes a tool to generate
classes which make up the AST (viz. expressions and statements).

This is purely a dev tool for playing with different language elements,
and as such is an exercise in producing write-only code. Using Kotlin
string templates, it might not be precisely worse than programming in
regular expressions... but the difference is not by much.

Additionally, the stuff.txt file contains interesting venues to pursue
at a later time.