Ditch visitor pattern
Got rid of visitors in favor of when pseudo-pattern-matching.
This removes the only (#visit) method from them, so all are
data classes now.

DESIGN NOTE:
Now, if (and only if) the problem of performance comes up,
and JMH or other checks put blame on type checking,
maybe I'll go and revert the change. The commit message
should helpfully pop up when searching for 'performance'
or 'design note'. Until then, I'm going for code readability
and reason-about-ability.